requirements.txt
# -----------------------------
# Core Django
# -----------------------------
Django>=5.0,<5.2
djangorestframework>=3.15

# -----------------------------
# MySQL Database
# -----------------------------
mysqlclient>=2.2.0

# -----------------------------
# Authentication & Security
# -----------------------------
python-dotenv>=1.0.1
django-cors-headers>=4.3.1

# -----------------------------
# File & Media Handling
# -----------------------------
Pillow>=10.0.0

# -----------------------------
# CSV & PDF Processing
# -----------------------------
pandas>=2.2.0
python-dateutil>=2.9.0
pdfplumber>=0.11.0
tabula-py>=2.9.0

# -----------------------------
# Email
# -----------------------------
django-anymail>=10.3

# -----------------------------
# OpenCV & Face Recognition
# -----------------------------
opencv-python>=4.9.0.80
face-recognition>=1.3.0
dlib>=19.24.2

# -----------------------------
# Utilities
# -----------------------------
numpy>=1.26.0
scipy>=1.12.0
requests>=2.31.0

# -----------------------------
# Deployment (optional)
# -----------------------------
gunicorn>=21.2.0
whitenoise>=6.6.0
